Revamp file structure to state-view
