--- src/components/containers/index.js
@@ -1,6 +1,0 @@
-/**
- * Barrel file for smart container components, which will be used to render
- * presentational components and handle state and dispatch mapping to connect 
- * to the Redux store. These will be incorporated within the main App component.
- */
-export * from "../containers";
