--- src/components/views/index.js
@@ -1,6 +1,0 @@
-/**
- * Barrel file for presentation components, which will be used to 
- * display the main JSX elements. These will be used in their corresponding 
- * smart containers.
- */
-export * from "../views";
