+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './app/App';
-import store from './store/store';
+import App from './view/App/App';
+import store from './state/shared/store/store';
 import { Provider } from 'react-redux';
 import * as serviceWorker from './serviceWorker';
