--- src/reducers/index.js
@@ -1,8 +1,0 @@
-/**
- * Barrel file for reducers, which will be combined and passed into the
- * Redux store we create. The aliases of reducers in this file will be 
- * assigned as the names of the keys in the Redux store, with the values 
- * being the respective individual reducers
- */
-export { default as counter } from "../store/utilities/counter";
-export { increment, decrement } from "../store/utilities/counter";
