+++ src/state/actions.js
@@ -1,0 +1,3 @@
+// Barrel file for exporting all actions
+
+export { increment, decrement } from "./counter";
