+++ src/state/counter/action_types.js
@@ -1,0 +1,5 @@
+// ACTION TYPES
+const INCREMENT = "INCREMENT";
+const DECREMENT = "DECREMENT";
+
+export default { INCREMENT, DECREMENT }
