+++ src/state/counter/index.js
@@ -1,0 +1,4 @@
+// Barrel file for exporting actions and reducers for counters
+
+export { default as counter } from "./reducers";
+export { increment, decrement } from "./actions";
