+++ src/state/reducers.js
@@ -1,0 +1,3 @@
+// Barrel file for exporting all reducers
+
+export { counter } from "./counter";
