+++ src/state/shared/store/store.js
@@ -5,7 +5,7 @@ import thunkMiddleware from "redux-thunk";
 import { composeWithDevTools } from "redux-devtools-extension";
 
 // Individual reducers altogether under an alias
-import * as reducers from '../reducers';
+import * as reducers from '../../reducers';
 
 // Construct our Redux store
 const rootReducer = combineReducers(reducers);
